new: Render layer consolidation and multi-texturing for OpenGL 3.3+
Firstly, the solid and cutout render layers are merged into one. This is
fairly trivial and doesn't come with any downsides, and the shader code
only needs to be slightly adjusted. It looks like Mojang was trying to
avoid the expensive alpha test/func, but in our shader code this is
basically free.

Then, a multi-texturing approach is used to allow consolidation of the mipped
and non-mipped render layers. The vertex format is extended to include a
fourth component (w) in the position attribute, which is used as the blend
factor between the mipped and non-mipped texture atlases. Each block layer
can specify whether or not is mipped during mesh generation, which controls
the value of the w-component in chunk meshing.

Multi-texturing requires the usage of sampler objects from OpenGL 3.3+, as
the same texture object cannot be bound with different texture parameters
normally. For hardware that doesn't support the OpenGL 3.3 profile, the
ARB_sampler_objects extension is identical and available on almost all
hardware that can run Minecraft.

With all these changes, this enables all non-translucent render passes
to be consolidated into one draw call per chunk, greatly speeding up
rendering performance by eliminating the render layer setup code and many
expensive state switches.

The overhead of needing to sample the same block atlas texture is probably
negligible, and is much faster than alternative techniques which involve
switching between sampler uniforms or branching.